Refactor S3OutputStreamImplementations to extend an abstract class with common code for creation of the OutputStream based on a temporary file
