+++ spring-cloud-aws-s3-parent/spring-cloud-aws-s3/src/main/java/io/awspring/cloud/s3/DiskBufferingS3OutputStream.java
@@ -15,23 +15,12 @@
  */
 package io.awspring.cloud.s3;
 
-import java.io.BufferedOutputStream;
-import java.io.File;
-import java.io.FileOutputStream;
-import java.io.IOException;
-import java.io.OutputStream;
-import java.security.DigestOutputStream;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-import java.util.Base64;
-import java.util.UUID;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.lang.Nullable;
-import org.springframework.util.Assert;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.PutObjectRequest;
+
+import java.io.IOException;
 
 /**
  * {@link S3OutputStream} implementation, that first uploads to a local file in tmp folder and then flushes file to S3.
@@ -43,40 +32,9 @@ import software.amazon.awssdk.services.s3.model.PutObjectRequest;
  * @author Maciej Walkowiak
  * @since 3.0
  */
-class DiskBufferingS3OutputStream extends S3OutputStream {
-
-	private static final Logger LOG = LoggerFactory.getLogger(DiskBufferingS3OutputStream.class);
-
-	private final Location location;
-
-	/**
-	 * The local file that will be uploaded when the stream is closed.
-	 */
-	private final File file;
+class DiskBufferingS3OutputStream extends BaseTempFileS3OutputStream {
 
 	private final S3Client s3Client;
-
-	/**
-	 * The outputstream to a local file where the file will be buffered until closed.
-	 */
-	private OutputStream localOutputStream;
-
-	/**
-	 * The MD5 hash of the file.
-	 */
-	@Nullable
-	private MessageDigest hash;
-
-	@Nullable
-	private final ObjectMetadata objectMetadata;
-
-	/**
-	 * Flag to indicate this stream has been closed, to ensure close is only done once.
-	 */
-	private boolean closed;
-
-	@Nullable
-	private final S3ObjectContentTypeResolver contentTypeResolver;
 
 	DiskBufferingS3OutputStream(Location location, S3Client s3Client, @Nullable ObjectMetadata objectMetadata)
 			throws IOException {
@@ -85,74 +43,12 @@ class DiskBufferingS3OutputStream extends S3OutputStream {
 
 	DiskBufferingS3OutputStream(Location location, S3Client client, @Nullable ObjectMetadata objectMetadata,
 			@Nullable S3ObjectContentTypeResolver contentTypeResolver) throws IOException {
-		Assert.notNull(location, "Location must not be null.");
-		this.location = location;
+		super(location,objectMetadata, contentTypeResolver);
 		this.s3Client = client;
-		this.objectMetadata = objectMetadata;
-		this.contentTypeResolver = contentTypeResolver;
-		this.file = File.createTempFile("DiskBufferingS3OutputStream", UUID.randomUUID().toString());
-		try {
-			hash = MessageDigest.getInstance("MD5");
-			localOutputStream = new BufferedOutputStream(new DigestOutputStream(new FileOutputStream(file), hash));
-		}
-		catch (NoSuchAlgorithmException e) {
-			LOG.warn("Algorithm not available for MD5 hash.", e);
-			hash = null;
-			localOutputStream = new BufferedOutputStream(new FileOutputStream(file));
-		}
-		closed = false;
 	}
 
 	@Override
-	public void write(int b) throws IOException {
-		localOutputStream.write(b);
+	protected void upload(PutObjectRequest putObjectRequest) {
+		s3Client.putObject(putObjectRequest, RequestBody.fromFile(file));
 	}
-
-	@Override
-	public void write(byte[] b) throws IOException {
-		localOutputStream.write(b, 0, b.length);
-	}
-
-	@Override
-	public void write(byte[] b, int off, int len) throws IOException {
-		localOutputStream.write(b, off, len);
-	}
-
-	@Override
-	public void flush() throws IOException {
-		localOutputStream.flush();
-	}
-
-	@Override
-	public void close() throws IOException {
-		if (closed) {
-			return;
-		}
-		localOutputStream.close();
-		closed = true;
-		try {
-			PutObjectRequest.Builder builder = PutObjectRequest.builder().bucket(location.getBucket())
-					.key(location.getObject()).contentLength(file.length());
-			if (objectMetadata != null) {
-				objectMetadata.apply(builder);
-			}
-			if (hash != null) {
-				String contentMD5 = new String(Base64.getEncoder().encode(hash.digest()));
-				builder = builder.contentMD5(contentMD5);
-			}
-			if (contentTypeResolver != null && (objectMetadata == null || objectMetadata.getContentType() == null)) {
-				String contentType = contentTypeResolver.resolveContentType(location.getObject());
-				if (contentType != null) {
-					builder.contentType(contentType);
-				}
-			}
-			s3Client.putObject(builder.build(), RequestBody.fromFile(file));
-			file.delete();
-		}
-		catch (Exception se) {
-			LOG.error(String.format("Failed to upload %s. Temporary file @%s", location.getObject(), file.getPath()));
-			throw new UploadFailedException(file.getPath(), se);
-		}
-	}
-
 }
